RAG-KDD-Cup/Meta KDD Cup CRAG/CRAG-MM 2025/09-Winning-Task-2/code/crag-mm/requirements.txt
# File transfer and loading utilities
hf_transfer
dotenv

# Core ML frameworks
torch #>2.6
transformers>=4.52.4
accelerate>=0.26.0
sentence-transformers
# vllm>=0.6.2 #,<0.8.0 # We need at least 0.6.2 to support LLaMA3.2-Vision. 

# Data processing and modeling
pandas
pydantic
datasets
Pillow
jinja2>=3.1.0
tokenizers

# API integrations
openai
cragmm-search-pipeline>=0.5.0
chromadb>=1.0.0

# Misc
loguru
rich
tqdm

# others
kagglehub
peft
bitsandbytes
matplotlib
seaborn
wandb